+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Improvements:
 * Core: Upgraded to CKEditor 3.6.4
 * Profiles: rewrote method for inserting/multiple settings.
 * Profiles: ask a display name in the register-step.
+* Profiles: redirect to login if the profile isn't logged in on a settings-page.
 
 Bugfixes: